+++ libs/shared/feature/i18n-nest/.eslintrc.json
