+++ libs/shared/feature/i18n-nest/jest.config.ts
