+++ libs/shared/feature/i18n-nest/project.json
