+++ libs/shared/feature/i18n-nest/tsconfig.lib.json
