+++ libs/shared/feature/i18n-nest/tsconfig.spec.json
