+++ libs/shared/feature/logger/.eslintrc.json
