+++ libs/shared/feature/logger/project.json
