+++ libs/shared/feature/logger/tsconfig.lib.json
