+++ libs/shared/feature/logger/tsconfig.spec.json
