+++ libs/shared/feature/mongo/.eslintrc.json
