+++ libs/shared/feature/mongo/project.json
