+++ libs/shared/feature/mongo/src/lib/base.schema.ts
