+++ libs/shared/feature/mongo/tsconfig.json
