+++ libs/shared/feature/mongo/tsconfig.lib.json
