+++ libs/shared/feature/mongo/tsconfig.spec.json
