+++ tsconfig.base.json
@@ -28,9 +28,11 @@
       "@towech-finance/desktop/toasts/error": ["libs/desktop/toasts/feature/error-handler/src"],
       "@towech-finance/desktop/toasts/tray": ["libs/desktop/toasts/feature/tray/src"],
       "@towech-finance/desktop/toasts/ui": ["libs/desktop/toasts/ui/src"],
-      "@towech-finance/shared/features/i18n-nest": ["libs/shared/features/i18n-nest/src"],
-      "@towech-finance/shared/features/logger": ["libs/shared/features/logger/src"],
-      "@towech-finance/shared/features/mongo": ["libs/shared/features/mongo/src"],
+      "@towech-finance/desktop/user/data-access": ["libs/desktop/user/data-access/src"],
+      "@towech-finance/desktop/user/guards": ["libs/desktop/user/feature/guards/src"],
+      "@towech-finance/shared/feature/i18n-nest": ["libs/shared/feature/i18n-nest/src"],
+      "@towech-finance/shared/feature/logger": ["libs/shared/feature/logger/src"],
+      "@towech-finance/shared/feature/mongo": ["libs/shared/feature/mongo/src"],
       "@towech-finance/shared/ui/button": ["libs/shared/ui/button/src"],
       "@towech-finance/shared/ui/checkbox": ["libs/shared/ui/checkbox/src"],
       "@towech-finance/shared/ui/input": ["libs/shared/ui/input/src"],
